+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased changes (2026-04-11)
 
-* fix: set conf timeout to cutoff before runner construction in wrappers [[details](https://github.com/cbboyan/solverpy/commit/) | [browse](https://github.com/cbboyan/solverpy/tree/)]
+* fix: set conf timeout to cutoff before runner construction in wrappers [[details](https://github.com/cbboyan/solverpy/commit/f83b017) | [browse](https://github.com/cbboyan/solverpy/tree/f83b017)]
 * fix: eprover domain defaults match E binary defaults; clean() strips them; -H moved to end of args [[details](https://github.com/cbboyan/solverpy/commit/999e6fc) | [browse](https://github.com/cbboyan/solverpy/tree/999e6fc)]
 * refactor: extract HoDomain from CoreDomain; parametric HeuristicDomain max_slots [[details](https://github.com/cbboyan/solverpy/commit/b06ffb0) | [browse](https://github.com/cbboyan/solverpy/tree/b06ffb0)]
 * refactor: replace old wrapper scripts with new grackle scripts [[details](https://github.com/cbboyan/solverpy/commit/ff8a124) | [browse](https://github.com/cbboyan/solverpy/tree/ff8a124)]